requirements.txt
flask
faiss-cpu
sentence-transformers
langchain-groq
langchain-community
python-dotenv
pypdf
unstructured
openpyxl
docx2txt
numpy
torch
